qmlls: fix crash when constructing for-loop
There is a qmlls crash (assertion was thrown in qqmldomastcreator when
constructing the for-loop) for the code I wrote for the blogpost. Added
the problematic function to tst_qmldomitem:forStatements.

Fix the if-conditions used in the macros to no run into the asserts.

The crash does not happen on dev, only on 6.6beta2.

Change-Id: Ie65f49cdbc444265a5bd604c2b4e00f7a3727bde
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>