QQuickAnimation: Check animationInstance for null in animationGroupDirty
On this line group->currentTime() defaults to 0 if animationInstance of
group is null. This happens to be the value the condition is checking
for and leads to a failed Q_ASSERT(animationInstance) in
animGroupPriv->restartFromCurrentLoop(). Add a check for nullptr to
cover this case.

Amends 0e69268b49f6eed71a7a3e6f16bb5098b572ef2b

Task-number: QTBUG-68404
Pick-to: 6.5 6.6
Change-Id: I12236be3b42fc6bd2d185943f9e0474655d8a4c9
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Inho Lee <inho.lee@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>