qmlls: find id's, not the item around it
When you use qmlls's 'go to definition' on an id, it jumps to the
object around it, instead of the id-binding. For example:

```
Item { id:me }
// go to definition for 'me' goes to 'Item' above instead of the 'me'
// after the id!
function f() { console.log(me)}
```

Retrieve the correct sourcelocation from the Dom, fix the tests,
and remove one unused helper method in tst_qmlls_modules.

Change-Id: Ic25ceb8399cbc27e65574f2c3c7e71249bcac8f6
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>