QtQml: unbreak unity-build-batch-size 100000
Exclude TUs that cause problems in a build where all of QtQml's .cpp
files end up in a single unity_0_cxx.cxx. This should ensure that the
build will forthwith not fail because someone added a new .cpp file in
the "wrong" position.

Of course, this is just a snapshot, with my configuration: Clang 15,
Ubuntu 20.04, -developer-build, C++23, -sctp, libc++, ...

Task-number: QTBUG-115140
Pick-to: 6.6 6.5
Change-Id: If3cb7c17548710dca3b6111b33a4d3ca70417728
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>