qmlls: Add support for range formatting
Implement a selection based code formatter. A range
formatter is a code formatting tool that formats only a selected portion
of the code rather than the entire file. We now don't check the validity
of dom document, instead we read the source code and obtain the token
status of the previous line of the selected range. Then, indenting line
writer perfoms the indentation based on the last status.

For now, it has a limited functionality that it can only perfom
reindentation and whitespace handling. We will improve that with
QTBUG-116139

Exclude newly added test files from qmlformat test cases since currently
qmldom cannot handle unicode characters in code.

Fixes: QTBUG-112978
Change-Id: Iabf392a7b2623cfee44d042e1c87b021ccc6e5e6
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>