Use pixel aligned value in Flickable ends check
When animating a flickable with pixelAlignment enabled, the vData and
hData move still contain raw floating values.

The viewport and content items will use the pixel aligned values. We
recheck updateBeginningEnd based on when the viewport changes to emit
the atBeginning/atEnd signals.

If the animations speeds are just right, we can have a frame where the
hData.move is not at the end, but the aligned version is. This will
perform the last check in a way that fails the test for being at the
end, and not update again.

This patch uses the same value for the position as the content, aligning
if needed.

Fixes: QTBUG-115081
Pick-to: 6.5 6.6
Change-Id: I58afbe8a79994507e891391cb8f2af7e22380a9b
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>