QmlCompiler: Do not coerce numbers to strings
We have QJSPrimitiveValue for this, and we can still coerce
QJSPrimitiveValue to string if we really need to. However, if you have
overloads of QString and QJSPrimitiveValue for some method, you really
want the QJSPrimitiveValue overload to be called if a number is passed.

Pick-to: 6.5 6.2
Task-number: QTBUG-111624
Change-Id: Icbc772065ba362994eaa8cb23025ed0bb95eee9b
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
(cherry picked from commit 2a998f2b043242df965bb1d21495fd5e299de322)