qmltc: fix GCC13 warnings "class X is implicitly friend with itself"
This happens because the generated code contains

    class X {
        ~~~~
    private:
        friend class X; // warning here
    };

Says GCC13 (e.g.):

    tests/auto/qml/qmltc/QmltcTests/.qmltc/qmltc_test_module/inlinecomponents.h:421:18: warning: class ‘QmltcTests::inlineComponents_IC0’ is implicitly friends with itself
      421 |     friend class inlineComponents_IC0;
          |                  ^~~~~~~~~~~~~~~~~~~~

It seems the intended check for documentRoot doesn't work in all
cases, so do the check on the class-name level.

There appears to be no -W flag to control this warning, therefore no
apparent way to suppress it, so we need to fix it.

Amends either b89a92053e1a4565d37f75831db1c6c90c21bce6 or
0990b892cab819f210120f8a9f9ca48522bde412. I didn't dig into it
further.

Pick-to: 6.6 6.5
Change-Id: I3fe653a398ea5b7a3e045fd3ea8dfb5d5c0f2e5c
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>