Dialogs: fix statically-built applications not running
When attemping to run a statically-built Qt Quick app that uses
QtQuick.Dialogs, users would get an error:

    qrc:/qt-project.org/imports/QtQuick/Dialogs/quickimpl/qml/MessageDialog.qml:
    module "QtQuick.Dialogs.quickimpl" plugin "qtquickdialogs2quickimplplugin"
    not found

This patch fixes these errors by listing QtQuick.Dialogs.quickimpl as a
dependency through the qt_internal_add_qml_module command.

This does introduce an unconditional tooling (e.g. qmlimportscanner)
dependency on the non-native Qt Quick fallback plugin for all users,
regardless of whether they are using a statically-built Qt or not.
However, this should be fine, and the decision of whether or not to use
the fallback is made at runtime anyway, so it may not be possible to
conditionally add the dependency at CMake configure time.

Fixes: QTBUG-105745
Pick-to: 6.2 6.5 6.6
Change-Id: If2e4c4346963e645bd8e5e9cc210c7b38c8616c0
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>