TextEdit/TextArea: don't deselect on release if IM just selected text
After 90d3fac73a10b9363fd34e6757cc730d7a0c086c we got a regression on
Android at least: if you long-press to select a word, selection handles
appear; then when you release, there's a QTouchEvent which is left
unhandled, and then a synth-mouse release. In this case, we don't want
to call setCursorPosition(), because it de-selects the selection that
was just made by the input method.

But the intention was to be able to set the cursor position (and also
deselect text) by tapping on a touchscreen on other platforms that don't
have text selection handles. (It seems to be consistent with Android
behavior: there you can also deselect and move the cursor by simply
tapping.) To avoid changing that behavior, we now set a flag if we get
a mouse press and then a QInputMethodEvent with a Selection attribute,
indicating that text got selected that way; and in that case, don't
call setCursorPosition().

Fixes: QTBUG-115004
Pick-to: 6.5
Change-Id: I7e362da584f6e1cb1509dcc0eff024b291ef5df3
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
Reviewed-by: Rami Potinkara <rami.potinkara@qt.io>
(cherry picked from commit c1d5e46ea79762bfcc070fabbf10ff7fdf184e01)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>