TreeViewDelegate: don't use deprecated API
TableView.modelIndex(column, row) was deprecated in Qt 6.4
(291d3ed0da) in favor of TableView.index(row, column). But the
TreeViewDelegates are still using the deprecated API for the
edit delegates (added in Qt 6.5, 21bf976a41), which will cause
warning messages to be printed.

This patch will rewrite the offending lines in the
TreeViewDelegates to use TableView.index(row, column)
instead.

Pick-to: 6.6 6.5
Task-number: QTBUG-109542
Change-Id: I41c0c18934493611663b759c8eb3b4fed2510cc6
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>