ListView: add extra note about reusing items
The current documentation leads the reader to think
that all items are always pooled as soon as they
are flicked out the viewport. Add a small note that
explains that this is not always the case.

Task-number: QTBUG-115317
Change-Id: I31bc7920ae7fa2477764f1f2ce33a29ef6825a3f
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit c0ab315b8582e39c95e0755e26424bb331a755d5)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>