TestModel: port away from Q_FOREACH
Replace the foreach loop with a call to qDeleteAll().

This is required, in the grand scheme of things, because the code is
in a header, and that makes any TU including it (all, in case of PCH)
incompatible with QT_NO_FOREACH.

This is also safe, however, because a) Node is non-polymorphic and b)
the Node destructor (the function containing the loop!) doesn't
attempt to unregister from its parent, so we know exactly what code is
being run here, and it's not modifying the container.

Task-number: QTBUG-115808
Change-Id: I5e434d55945fa36b318427a4ea528ac4991422c8
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>