QQuickTableView: let the subclass decide which model to use for selections
When a subclass that has a proxy model is in use, it's not given
if the selection model should be using the application model
or the proxy model to store selections. Rather, this is subclass
specific. In case of TreeView, the application model should
be used. In case of HeaderView, the proxy model should be
used. Until now, we always used the application model, which
would fail if trying to select cells in HeaderView. The reason
for the latter is that the proxy model in HeaderView doesn't
share any common model items with the application model, and
therefore cannot use the application model for selections.

Fixes: QTBUG-115227
Pick-to: 6.5 6.2
Change-Id: I39f0687e0396392039234787fda0df6d7d4d7da4
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 82d0d8321a02f46732489b5fedc8e77e56523c3d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>