Doc: Add note to the QML Shortcut sequence property
When using a standard key for the sequence property, the key may resolve
to multiple sequences, depending on the platform. This results in a
warning, which tells us that only the first sequence gets assigned
to the shortcut and that one should use the sequences property
instead. Add a note about this in the documentation as well.

Also, change the top-most example to use sequences instead sequence.

Fixes: QTBUG-115947
Pick-to: 6.6 6.5
Change-Id: I9f65e0e9ac9473f536616fc3979965bc4702f6b8
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>