Merge variables for tracking component completion for QQuickWindow
The QQuickWindowPrivate::componentCompleted member was not used for
anything, while the QQuickWindowQmlImplPrivate::complete member was
doing the completion tracking.

The variable is now named componentComplete, to match QQuickItem,
and lives in the QQuickWindowPrivate base class.

As seems customary for other classes, the variable is default
initialized to true, and then reset to false in classBegin(),
and then to true in the componentComplete() callback.

Pick-to: 6.6
Change-Id: I77e5864d1c21deaf87d9098a149d590cfc160e4d
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>