macOS: add Switch
This adds a QML-based implementation of Switch for the macOS style.
Switch doesn't exist in widgets, so there is no QStyle support for
this control.

__isDefaultDelegate was renamed to __focusFrameRadius and turned into
a qreal so that QML types can have control over the size of the focus
frame radius. No QML types were using __isDefaultDelegate.

Task-number: QTBUG-115165
Change-Id: I4f6c961ab809ba7c5c9bfccb3218d33316dd72c2
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>