QML: Load the implicit import before resolving scripts
Scripts listed in the qmldir of the implicit import have to be made
available as references in the context.

Since we now also load singletons from the implicit import and check
them for errors, the test that deals with errors in singletons in the
implicit import has to be adapted.

Change-Id: Id324b2d9f646d9f48cf46e1c9dd89a0701fcea7e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>