SwipeView: ensure items fill the bounds
Taking the translucentPages component from
SwipeView::test_initialPositions into a Qt Quick app and running it
showed that the Text items weren't actually filling the SwipeView. The
documentation states:

"SwipeView is populated with a set of pages. One page is visible at a
time."

The implication here being that each item fills the SwipeView, hence only
one being visible at a time.

It also states that SwipeView manages the geometry of items:

"Note: SwipeView takes over the geometry management of items added to
the view. Using anchors on the items is not supported, and any width or
height assignment will be overridden by the view. Notice that this only
applies to the root of the item. Specifying width and height, or using
anchors for its children works as expected."

So, this patch corrects the behavior and updates the test.

Change-Id: I3425d4d028743f440f003e86ba009a2da7cca9cb
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 08b23023109de2cf91c0badf3a2d22a13cb09f40)