CMake: Add the Svg package to OPTIONAL_COMPONENTS
It is an optional dependency of the module and some examples and tests
depend on it.

Since this patch enables the Svg package on the CI, a previously
unexecuted test has begun failing on the Android platform. This seems to
be a known issue, thus add this particular test case to the BLACKLIST
file.

Change-Id: I46e00308a81189c80c7d788920908a884d05663d
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>