Support targets for DEPENDENCIES/IMPORTS in qt_add_qml_module
Targets can be marked by prefixing them with TARGET.
This is only enabled when a new policy (QTP0005) is set to avoid
breaking existing code which might have used TARGET as a URI. Moreover,
we still support passing plain URIs, as not every module might be backed
by a CMake target.

[ChangeLog][QtQml][CMake API] qt_add_qml_module now supports passing
targets to the DEPENDENCIES and IMPORT keywords when prefixed with
TARGET. This requires opting into Qt's CMake policy QTP0005.

Task-number: QTBUG-114969
Change-Id: Ia80504f0ddaa22d1cf9fdb3a2b8680f5d578890f
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>