Remove undocumented HDR mode getter/setter from QQuickGr.Conf
Added in 6.6 as \internal but concluded that it is too
limited in this form.

Added originally for consistency, so that at least
QSG_RHI_HDR=auto (where said "auto" mode is also
new in 6.6) has a C++ equivalent, as is the custom
with QSG_* and QQuickGraphicsConfiguration, but it
is becoming clear that an "auto" option makes no sense
in practice.

A better and more holistic long-term plan is to handle
this in the Gui layer (e.g. QSurfaceFormat) eventually.

In the meantime, the env.var. (QSG_RHI_HDR) continues to
function like before with explicit format keys ("scrgb",
"hdr10").

Pick-to: 6.6
Change-Id: Iea667d7e46ffafa33fdff342d4ed749bdb1cfd4b
Reviewed-by: Christian Strømme <christian.stromme@qt.io>