QV4::QObjectWrapper: Use the object's actual meta type
Otherwise, a derived type's methods will not be found.

In our particular case we tried to call a method of 'B*' on a
QML attached property declared as 'A*'.

This restores the previous behavior of using the object()
for the meta type except for when it's a value type wrapper.

Amends commit 63b622d5908ec2960ce5dfa301e9d3fd4d92fdb4.

Pick-to: 6.6 6.5 6.2
Change-Id: I08b9f4b97a58c15fdc3703dd3c5d927cd1beb3ce
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Semih Yavuz <semih.yavuz@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>