StyleGenerator: Improve how we export transitions
Since the Figma templates have changed structure, the transition
information is now not anymore stored in objects of "type"
"COMPONENT", but rather of type "INSTANCE". To be more accurate,
we need to make sure that this instance represents a VARIANT, and
more specifically, the state we're looking for.
This makes it a bit more flexible as well.

Change-Id: I66a8fba98487dad0076a1fa7d194d5225dd0bcab
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>