qmlls: fix 'qmldir contains dependency...' warning
In qmlls, there is a warning printed out all the time that goes
'qmldir contains dependency with auto keyword'. It seems to indicate
that QmlDom does not support depends-commands with 'auto' versions, and
instead chooses the latest version.

Rewrite the warning to be more understandable and hide it from the users because
'auto' keywords in depends command happen quite often in the QtQuick
modules. This is nothing that the user can fix or should worry about, so
move the warning from qWarning to qCDebug().

Change-Id: I72d2ba878735a61a921d998874fc5f0f9855fbc5
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>