macOS: correct PM_*FocusFrameRadius values
Until the patch that implements Switch (follow-up to this change),
__isDefaultDelegate was never actually read. That property allows
controls that are implemented purely in QML to specify their own focus
frame radius, and will be renamed to __focusFrameRadius in the
follow-up change.

However, if we simply pass the control's desired radius to
moveToItem(), the effective focus frame radius will be too small. This
is because the PM_*FocusFrameRadius values from
QMacStyle::pixelMetric() already account for the frame size. To make
the focus frame have the correct size for both QStyle and non-QStyle
controls, we adjust the M_*FocusFrameRadius values and instead account
for FocusFrame's frame size in QML.

This patch also increases the frame size to 3, as that better matches
what I see natively, and also because QMacStyle::pixelMetric can only
return an int anyway.

Task-number: QTBUG-115165
Change-Id: Ic527fa64ac1a4cd08b0b99a7426fca64e757eb76
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>