StyleGenerator: add StyleImage
Add a new component called StyleImage. This item will contain
an image that can have both radii and shadows. Information about
the latter is available from the config file as offsets and insets
(respectivly), and will be used by StyleImage to ensure that the
child image scales correctly, and is also positioned correctly.

The point is that shadows, blur, and other effects that extend
outside the shape drawn by the image, should only be graphical, and
not be part the image' geometry. By "hiding" this information, users
of StyleImage doesn't need to take into accunt whether the image has
shadows or not when calculating its position inside the control.

StyleImage also provides a "horizontal" property that can be
set to false in order to pretend that the loaded image was
verical by flipping the image 90 degrees.

Change-Id: I80e0a34c1550fcf8115949b1b9e3c98f7d319718
Reviewed-by: Doris Verria <doris.verria@qt.io>