QQuickAbstractButton: accept only left mouse button
Normally, Button should only handle the left mouse button.
To do this, we even call setAcceptedMouseButtons(Qt::LeftButton) in
the QQuickAbstractButton's constructor.

But since 70113ef66b245e58ce12e2a9d26268e2eaeb3a42
QQuickItem's acceptedMouseButtons is overridden with Qt::AllButton
every time when any handler is set on the Item.
This caused a bug where the Button responds to the right mouse button
when, for instance, a HoverHandler is set on it.

As a solution, we ignore a mouse press event explicitly in
QQuickAbstractButton::mousePressEvent() if the left button is not pressed.

Fixes: QTBUG-116289
Pick-to: 6.5
Change-Id: I6f30da38fd7be274eb70f2e761789a9f49fefbff
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 7211543ecffb8ebe12c943135d8fcd8761f7c4a9)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>